@@ -3,25 +3,16 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.2.35",
-    "@fortawesome/free-solid-svg-icons": "^5.15.3",
-    "@fortawesome/react-fontawesome": "^0.1.14",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.6",
-    "@testing-library/user-event": "^12.8.3",
-    "@types/jest": "^26.0.23",
-    "@types/node": "^12.20.12",
-    "@types/react": "^17.0.5",
-    "@types/react-dom": "^17.0.3",
-    "axios": "^0.21.1",
-    "chart.js": "^3.2.1",
-    "node-sass": "^5.0.0",
-    "react": "^17.0.2",
-    "react-chartjs-2": "^3.0.3",
-    "react-dom": "^17.0.2",
-    "react-scripts": "4.0.3",
-    "typescript": "^4.2.4",
-    "web-vitals": "^1.1.2"
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.18",
+    "axios": "^0.27.2",
+    "chart.js": "^3.7.1",
+    "react": "^18.1.0",
+    "react-chartjs-2": "^4.1.0",
+    "react-dom": "^18.1.0",
+    "react-scripts": "^5.0.1",
+    "sass": "^1.51.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -49,6 +40,9 @@
   },
   "devDependencies": {
     "@types/chart.js": "^2.9.32",
-    "eslint-plugin-react-hooks": "^4.2.0"
+    "@types/react": "^18.0.8",
+    "@types/react-dom": "^18.0.3",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "typescript": "^4.6.4"
   }
 }
